requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv
gunicorn==20.1.0
